risolti problemi importazione famiglie, creato nuovo formato excell per l'importazione di spine
